+++ tests/callbacks_test.cpp
@@ -152,6 +152,93 @@ void CallbacksFullFunctionTest(OptimizerType& optimizer,
   REQUIRE(cb.calledStepTaken == calledStepTaken);
 }
 
+template<typename OptimizerType>
+void CallbacksFullMultiobjectiveFunctionTest(OptimizerType& optimizer,
+                                             bool calledEvaluate,
+                                             bool calledGradient,
+                                             bool calledBeginEpoch,
+                                             bool calledEndEpoch,
+                                             bool calledBeginOptimization,
+                                             bool calledEndOptimization,
+                                             bool calledEvaluateConstraint,
+                                             bool calledGradientConstraint,
+                                             bool calledStepTaken)
+{
+  SchafferFunctionN1<arma::mat> SCH;
+
+  typedef decltype(SCH.objectiveA) ObjectiveTypeA;
+  typedef decltype(SCH.objectiveB) ObjectiveTypeB;
+
+  CompleteCallbackTestFunction cb;
+
+  arma::mat coordinates = SCH.GetInitialPoint();
+  std::tuple<ObjectiveTypeA, ObjectiveTypeB> objectives = SCH.GetObjectives();
+
+  optimizer.Optimize(objectives, coordinates, cb);
+
+  REQUIRE(cb.calledEvaluate == calledEvaluate);
+  REQUIRE(cb.calledGradient == calledGradient);
+  REQUIRE(cb.calledBeginEpoch == calledBeginEpoch);
+  REQUIRE(cb.calledEndEpoch == calledEndEpoch);
+  REQUIRE(cb.calledBeginOptimization == calledBeginOptimization);
+  REQUIRE(cb.calledEndOptimization == calledEndOptimization);
+  REQUIRE(cb.calledEvaluateConstraint == calledEvaluateConstraint);
+  REQUIRE(cb.calledGradientConstraint == calledGradientConstraint);
+  REQUIRE(cb.calledStepTaken == calledStepTaken);
+}
+
+template<typename OptimizerType>
+void EarlyStopCallbacksLambdaFunctionTest(OptimizerType& optimizer)
+{
+  arma::mat data, testData, shuffledData;
+  arma::Row<size_t> responses, testResponses, shuffledResponses;
+
+  LogisticRegressionTestData(data, testData, shuffledData,
+      responses, testResponses, shuffledResponses);
+
+  LogisticRegression<> lr(shuffledData, shuffledResponses, 0.5);
+  arma::mat coordinates = lr.GetInitialPoint();
+
+  EarlyStopAtMinLoss cb(
+      [&](const arma::mat& /* coordinates */)
+      {
+        return lr.ComputeAccuracy(testData, testResponses,
+                                  coordinates);
+      });
+
+  optimizer.Optimize(lr, coordinates, cb);
+}
+
+TEST_CASE("EarlyStopAtMinLossLambdaCallbackTest", "[CallbacksTest]")
+{
+  SMORMS3 smorms3;
+  EarlyStopCallbacksLambdaFunctionTest(smorms3);
+}
+
+TEST_CASE("EarlyStopAtMinLossCustomLambdaTest", "[CallbacksTest]")
+{
+  // Use the 50-dimensional Rosenbrock function.
+  GeneralizedRosenbrockFunction f(50);
+  // Start at some really large point.
+  arma::mat coordinates = f.GetInitialPoint();
+  coordinates.fill(100.0);
+
+  EarlyStopAtMinLoss cb(
+      [&](const arma::mat& coordinates)
+      {
+        // Terminate if any coordinate has a value less than 10.
+        double minValue = arma::abs(coordinates).min();
+        return (minValue < 10.0) ?
+          std::numeric_limits<double>::max() : minValue;
+      });
+
+  SMORMS3 smorms3;
+  smorms3.Optimize(f, coordinates, cb);
+
+  // Make sure that we did not get to the optimum.
+  for (size_t i = 0; i < coordinates.n_elem; ++i)
+    REQUIRE(std::abs(coordinates[i]) >= 3.0);
+}
 
 /**
  * Make sure we invoke all callbacks (AdaBound).
@@ -282,6 +369,18 @@ TEST_CASE("KatyushaCallbacksFullFunctionTest", "[CallbacksTest]")
   Katyusha optimizer(1.0, 10.0, 1, 3, 0, 1e-10, true);
   CallbacksFullFunctionTest(optimizer, true, true, false, false, true, true,
       false, false, true);
+}
+
+/**
+ * Make sure we invoke all callbacks (NSGA2).
+ */
+TEST_CASE("NSGA2CallbacksFullFunctionTest", "[CallbackTest]")
+{
+  arma::vec lowerBound = {-1000};
+  arma::vec upperBound = {1000};
+  NSGA2 optimizer(20, 5000, 0.5, 0.5, 1e-3, 1e-6, lowerBound, upperBound);
+  CallbacksFullMultiobjectiveFunctionTest(optimizer, false, false, false, false,
+      true, true, false, false, true);
 }
 
 /**
@@ -550,12 +649,12 @@ TEST_CASE("TimerStopCallbackTest", "[CallbacksTest]")
 {
   SGDTestFunction f;
   arma::mat coordinates = f.GetInitialPoint();
-  
+
   // Instantiate the optimizer with a number of iterations that will take a
   // long time to finish.
   Adam opt(0.5, 2, 0.7, 0.999, 1e-8, 2000000000, -100, false);
   arma::wall_clock timer;
-  
+
   timer.tic();
   // The optimization process should return in one second.
   opt.Optimize(f, coordinates, TimerStop(0.5));